unicomp/mini_m: move embedded numpad layer and custom keycode into keyboard.c
